// packages/cross-chain/src/lib/generated/v1/tx.ts
export enum vmType {
  EVM = 0,
  SVM = 1,
  MOVE_VM = 2,
  WASM_VM = 3,
  CAIRO_VM = 4,
  OTHER_VM = 5,
  UNRECOGNIZED = -1,
}
// ...
export interface AccountId {
  namespace: string;
  chainId: string;
  ownerKey: string;
  vmType: vmType;
}

export interface MsgDeployNMSC {
  signer: string;
  accountId: AccountId | undefined;
  txHash: string;
}

export interface MsgMintPush {
  signer: string;
  /** account_id is the identifier of the crosschain owner account */
  accountId:
    | AccountId
    | undefined;
  /** tx_hash is the hash of the transaction in which user locked the tokens */
  txHash: string;
}

export interface CrossChainPayload {
  /** EVM address as hex string (0x...) */
  target: string;
  /** Amount in wei as string (uint256) */
  value: string;
  /** ABI-encoded calldata */
  data: string;
  /** uint256 as string */
  gasLimit: string;
  /** uint256 as string */
  maxFeePerGas: string;
  /** uint256 as string */
  maxPriorityFeePerGas: string;
  /** uint256 as string */
  nonce: string;
  /** uint256 as string */
  deadline: string;
}
// ...
export const AccountId: MessageFns<AccountId> = {
  encode(message: AccountId, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.namespace !== "") {
      writer.uint32(10).string(message.namespace);
    }
    if (message.chainId !== "") {
      writer.uint32(18).string(message.chainId);
    }
    if (message.ownerKey !== "") {
      writer.uint32(26).string(message.ownerKey);
    }
    if (message.vmType !== 0) {
      writer.uint32(32).int32(message.vmType);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AccountId {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAccountId();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.namespace = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.chainId = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.ownerKey = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.vmType = reader.int32() as any;
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AccountId {
    return {
      namespace: isSet(object.namespace) ? globalThis.String(object.namespace) : "",
      chainId: isSet(object.chainId) ? globalThis.String(object.chainId) : "",
      ownerKey: isSet(object.ownerKey) ? globalThis.String(object.ownerKey) : "",
      vmType: isSet(object.vmType) ? vmTypeFromJSON(object.vmType) : 0,
    };
  },

  toJSON(message: AccountId): unknown {
    const obj: any = {};
    if (message.namespace !== "") {
      obj.namespace = message.namespace;
    }
    if (message.chainId !== "") {
      obj.chainId = message.chainId;
    }
    if (message.ownerKey !== "") {
      obj.ownerKey = message.ownerKey;
    }
    if (message.vmType !== 0) {
      obj.vmType = vmTypeToJSON(message.vmType);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<AccountId>, I>>(base?: I): AccountId {
    return AccountId.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AccountId>, I>>(object: I): AccountId {
    const message = createBaseAccountId();
    message.namespace = object.namespace ?? "";
    message.chainId = object.chainId ?? "";
    message.ownerKey = object.ownerKey ?? "";
    message.vmType = object.vmType ?? 0;
    return message;
  },
};
// ...
export const CrossChainPayload: MessageFns<CrossChainPayload> = {
  encode(message: CrossChainPayload, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.target !== "") {
      writer.uint32(10).string(message.target);
    }
    if (message.value !== "") {
      writer.uint32(18).string(message.value);
    }
    if (message.data !== "") {
      writer.uint32(26).string(message.data);
    }
    if (message.gasLimit !== "") {
      writer.uint32(34).string(message.gasLimit);
    }
    if (message.maxFeePerGas !== "") {
      writer.uint32(42).string(message.maxFeePerGas);
    }
    if (message.maxPriorityFeePerGas !== "") {
      writer.uint32(50).string(message.maxPriorityFeePerGas);
    }
    if (message.nonce !== "") {
      writer.uint32(58).string(message.nonce);
    }
    if (message.deadline !== "") {
      writer.uint32(66).string(message.deadline);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CrossChainPayload {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCrossChainPayload();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.target = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.value = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.data = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.gasLimit = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.maxFeePerGas = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.maxPriorityFeePerGas = reader.string();
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.nonce = reader.string();
          continue;
        }
        case 8: {
          if (tag !== 66) {
            break;
          }

          message.deadline = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CrossChainPayload {
    return {
      target: isSet(object.target) ? globalThis.String(object.target) : "",
      value: isSet(object.value) ? globalThis.String(object.value) : "",
      data: isSet(object.data) ? globalThis.String(object.data) : "",
      gasLimit: isSet(object.gasLimit) ? globalThis.String(object.gasLimit) : "",
      maxFeePerGas: isSet(object.maxFeePerGas) ? globalThis.String(object.maxFeePerGas) : "",
      maxPriorityFeePerGas: isSet(object.maxPriorityFeePerGas) ? globalThis.String(object.maxPriorityFeePerGas) : "",
      nonce: isSet(object.nonce) ? globalThis.String(object.nonce) : "",
      deadline: isSet(object.deadline) ? globalThis.String(object.deadline) : "",
    };
  },

  toJSON(message: CrossChainPayload): unknown {
    const obj: any = {};
    if (message.target !== "") {
      obj.target = message.target;
    }
    if (message.value !== "") {
      obj.value = message.value;
    }
    if (message.data !== "") {
      obj.data = message.data;
    }
    if (message.gasLimit !== "") {
      obj.gasLimit = message.gasLimit;
    }
    if (message.maxFeePerGas !== "") {
      obj.maxFeePerGas = message.maxFeePerGas;
    }
    if (message.maxPriorityFeePerGas !== "") {
      obj.maxPriorityFeePerGas = message.maxPriorityFeePerGas;
    }
    if (message.nonce !== "") {
      obj.nonce = message.nonce;
    }
    if (message.deadline !== "") {
      obj.deadline = message.deadline;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CrossChainPayload>, I>>(base?: I): CrossChainPayload {
    return CrossChainPayload.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CrossChainPayload>, I>>(object: I): CrossChainPayload {
    const message = createBaseCrossChainPayload();
    message.target = object.target ?? "";
    message.value = object.value ?? "";
    message.data = object.data ?? "";
    message.gasLimit = object.gasLimit ?? "";
    message.maxFeePerGas = object.maxFeePerGas ?? "";
    message.maxPriorityFeePerGas = object.maxPriorityFeePerGas ?? "";
    message.nonce = object.nonce ?? "";
    message.deadline = object.deadline ?? "";
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}